src/requirements.txt
asgiref==3.7.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
paho-mqtt==1.5.1
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
